backend/backend_llm_implemented_v0/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
motor==3.6.0
pydantic==2.9.2
pydantic[email]==2.9.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3

openai>=1.40.0
orjson>=3.10.7
numpy>=1.26.4
rank-bm25>=0.2.2
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0.post1
pypdf>=4.2.0
scikit-learn>=1.3.2
tqdm>=4.66.4
